MemberFooter を MemberCard に直書き
MemberCard 専用になった解除ボタン (MemberFooter) を MemberCard 内に
インライン化し、member-footer.tsx を削除。useDiscardMember フックは
member-list-hook.ts のまま利用。
